action renames for easier testing, basic refactoring of image to text service complete
