+++ .github/workflows/pro-image-to-text-test.yml
@@ -1,4 +1,4 @@
-name: python tests for meme search pro image to text generator
+name: image to text generator tests for pro app
 
 on:
   pull_request:
@@ -7,8 +7,8 @@ on:
       - "meme_search_pro/image_to_text_generator/**"
 
 jobs:
-  ruff:
-    name: lint with ruff
+  image_to_text_lint:
+    name: image_to_text_lint
     runs-on: ubuntu-latest
     timeout-minutes: 3
     steps:
@@ -19,24 +19,24 @@ jobs:
           args: "check --fix"
           config: .ruff.toml
           working-directory: ./meme_search_pro/image_to_text_generator/app
-  # test:
-  #   name: run pytest
-  #   runs-on: ubuntu-latest
-  #   timeout-minutes: 5
-  #   steps:
-  #     - name: Checkout code
-  #       uses: actions/checkout@v2
-  #     - name: Set up Python
-  #       uses: actions/setup-python@v2
-  #       with:
-  #         python-version: "3.12"
-  #     - name: Install dependencies
-  #       working-directory: ./meme_search_pro/image_to_text_generator/app
-  #       run: |
-  #         python -m pip install --upgrade pip
-  #         pip install -r requirements.test
-  #         pip install -r requirements.txt
-  #     - name: Run pytest
-  #       working-directory: ./meme_search_pro/image_to_text_generator/app
-  #       run: |
-  #         PYTHONPATH=. python3.12 -m pytest tests/
+  image_to_text_tests:
+    name: run pytest
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.12"
+      - name: Install dependencies
+        working-directory: ./meme_search_pro/image_to_text_generator
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.test
+          pip install -r requirements.txt
+      - name: Run pytest
+        working-directory: ./meme_search_pro/image_to_text_generator
+        run: |
+          PYTHONPATH=. python3.12 -m pytest tests/
