+++ .github/workflows/standard-app-build.yml
@@ -7,7 +7,7 @@ on:
       - "v*"
 
 jobs:
-  Build_And_Push:
+  standard_app_build:
     runs-on: ubuntu-22.04
     permissions:
       contents: read
