+++ .github/workflows/standard-app-test.yml
@@ -8,7 +8,7 @@ on:
       - "meme_search/meme_search_app/**"
 
 jobs:
-  ruff:
+  standard_app_lint:
     name: lint with ruff
     runs-on: ubuntu-latest
     timeout-minutes: 3
@@ -30,7 +30,7 @@ jobs:
         run: |
           ruff format --check --config ./meme_search/.ruff.toml ./meme_search/meme_search_app
 
-  test:
+  standard_app_tests:
     name: run pytest
     runs-on: ubuntu-latest
     timeout-minutes: 5
